将费用相关字段类型从Double更改为BigDecimal，以提高计算精度；更新相关方法以支持BigDecimal类型，优化CA代币余额检查和更新逻辑，提升整体系统的准确性和用户体验。

+++ src/main/java/online/bingzi/aetherbot/service/AiModelService.java
@@ -3,6 +3,7 @@ package online.bingzi.aetherbot.service;
 import online.bingzi.aetherbot.entity.AiModel;
 import online.bingzi.aetherbot.enums.ModelStatus;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AiModelService {
@@ -38,8 +39,8 @@ public interface AiModelService {
      * @param description                     模型描述
      * @return 创建的AI模型实体
      */
-    AiModel createModel(String name, Double promptCostPerThousandTokens,
-                        Double completionCostPerThousandTokens, String description);
+    AiModel createModel(String name, BigDecimal promptCostPerThousandTokens,
+                        BigDecimal completionCostPerThousandTokens, String description);
 
     /**
      * 更新AI模型状态
